requirements.txt
certifi==2020.6.20
chardet==3.0.4
ftfy==5.8
idna==2.10
mailchecker==3.3.15
pathlib2==2.3.5
phonenumbers==8.12.11
python-benedict==0.22.0
python-dateutil==2.8.1
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.11
wcwidth==0.2.5
xmltodict==0.12.0
